+++ VTOS.Application/Features/Users/DTOs/AddParentBankAccountRequest.cs
@@ -1,0 +1,8 @@
+namespace VTOS.Application.Features.Users.DTOs;
+
+public record AddParentBankAccountRequest(
+    string BankName,
+    string? BankCode,
+    string AccountNumber,
+    string AccountHolderName,
+    bool IsDefault = false);
